+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Highlights are marked with a pancake 🥞
 - Node API improvements [#1061](https://github.com/p2panda/p2panda/pull/1061)
 - Minor store improvements [#1068](https://github.com/p2panda/p2panda/pull/1068)
 - Use Borrow to express required args in ingest [#1078](https://github.com/p2panda/p2panda/pull/1078)
+- Use MockInstant for Timestamp when running tests [#1081](https://github.com/p2panda/p2panda/pull/1081)
 
 ### Fixed